FIN-Persona/app/requirements.txt
pandas
numpy
scikit-learn
openai
prometheus-client
python-dotenv